Add @pattern as a custom alias for DOxygen comments.
Migrated to clang and tried to use the modules C20 feature.  It's just too new and unstable, so I'm saving it here and then I'll unwind it, go back to gcc and make header files.

Created our first (prototype) test.  We need either modules or header files, so we have to make that work before we can do unit tests.