Refactor CommodityType.hpp so the only file it needs to be in is Commodity.  It's internal to Commodity, so let's keep it that way.
Refactor Commodity.hpp so that it can be in any/every class in Empire.

Refactor CommodityType's members to const rather than constexpr.  Because the array is created as a static, and has to be built
at runtime, we can't resolve the contents at compile time.

Build out the flyweight getters in Commodity.

Improve validation.

Set CommodityType and CommodityTypes to final.